refactor(client): use the new members data and restructure some file structure
