+++ src/client/features/Members/template.tsx
@@ -9,9 +9,9 @@ import { commonStyles } from 'client/styles/tokens';
 import { getMembersUrl, getMemberUrl } from 'client/utils/urls';
 import { useTranslations } from 'client/hooks/useTranslations';
 import { useIntl } from 'client/hooks/useIntl';
-import { MembersPageProps } from 'pages/members';
 import { useLanguageContext } from 'client/store/language/hook';
 import { PageHelmet } from 'client/layouts/PageHelmet';
+import { MembersPageProps } from 'client/features/Members/container';
 
 export const MembersPage: React.FC<MembersPageProps> = props => {
   const { currentFilter, memberGroupsByJoin } = props;
